requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.0
gunicorn==20.1.0
djangorestframework==3.13.1
psycopg2==2.8.4
Pygments==2.3.1
